+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Equipment
+- Filters for equipment
 
 ## [2.6.2] - 2019-04-28
 ### Fixed